requirements.txt
fastapi
netcdf4
xarray
rioxarray
shapely
pillow
uvicorn